+++ ownership-chain/node/src/chain_spec/klaos.rs
@@ -1,51 +1,16 @@
-use super::Extensions;
+use super::{endowed_accounts, get_collator_keys_from_seed, Extensions, SAFE_XCM_VERSION};
 use cumulus_primitives_core::ParaId;
 use fp_evm::GenesisAccount;
 use hex_literal::hex;
 use klaos_ownership_runtime::{AccountId, AuraId, Precompiles, REVERT_BYTECODE};
 use sc_service::ChainType;
-use sp_core::{Pair, Public, H160, U256};
+use sp_core::{H160, U256};
 use sp_runtime::traits::Zero;
 use std::{collections::BTreeMap, str::FromStr};
-
-/// List of endowed accounts.
-fn endowed_accounts() -> Vec<AccountId> {
-	vec![
-		// ALITH
-		hex!("f24FF3a9CF04c71Dbc94D0b566f7A27B94566cac").into(),
-		// BALTATHAR
-		hex!("3Cd0A705a2DC65e5b1E1205896BaA2be8A07c6e0").into(),
-		// CHARLETH
-		hex!("798d4Ba9baf0064Ec19eB4F0a1a45785ae9D6DFc").into(),
-		// DOROTHY
-		hex!("773539d4Ac0e786233D90A233654ccEE26a613D9").into(),
-		// ETHAN
-		hex!("Ff64d3F6efE2317EE2807d223a0Bdc4c0c49dfDB").into(),
-		// FAITH
-		hex!("C0F0f4ab324C46e55D02D0033343B4Be8A55532d").into(),
-	]
-}
 
 /// Specialized `ChainSpec` for the normal parachain runtime.
 pub type ChainSpec =
 	sc_service::GenericChainSpec<klaos_ownership_runtime::RuntimeGenesisConfig, Extensions>;
-
-/// The default XCM version to set in genesis config.
-const SAFE_XCM_VERSION: u32 = staging_xcm::prelude::XCM_VERSION;
-
-/// Helper function to generate a crypto pair from seed
-pub fn get_from_seed<TPublic: Public>(seed: &str) -> <TPublic::Pair as Pair>::Public {
-	TPublic::Pair::from_string(&format!("//{seed}"), None)
-		.expect("static values are valid; qed")
-		.public()
-}
-
-/// Generate collator keys from seed.
-///
-/// This function's return type must always match the session keys of the chain in tuple format.
-pub fn get_collator_keys_from_seed(seed: &str) -> AuraId {
-	get_from_seed::<AuraId>(seed)
-}
 
 /// Generate the session keys from individual elements.
 ///
