+++ ownership-chain/primitives/Cargo.toml
@@ -12,6 +12,7 @@ frame-support = { workspace = true }
 frame-system = { workspace = true }
 sp-core = { workspace = true }
 sp-runtime = { workspace = true }
+sp-consensus-aura = { workspace = true }
 
 # EVM
 fp-account = { workspace = true, features = ["serde"] }
@@ -24,4 +25,5 @@ std = [
 	"sp-core/std",
 	"sp-runtime/std",
 	"fp-account/std",
+	"sp-consensus-aura/std",
 ]
